+++ Changelog.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Apply log module to keep requested data and save to log file. Completed on 2025-10-13
 - Apply the sessionMiddleware in deskelectronjs. Completed on 2025-10-15
+- Add aes-256 encrypt and decrypt function to utils. Completed on 2025-10-16
 
 ### Changed